update django split different api to different django apps
