+++ MLabHubdjango/MLabHubdjango/urls.py
@@ -6,7 +6,6 @@ from django.contrib import admin
 urlpatterns = [
     path('api-auth/', include('rest_framework.urls')),
     path('api/', include('api.urls')),
-    path('lab/', include('lab.urls')),
     path('', views.index, name='index'),
     path('admin/', admin.site.urls)
 ]
