+++ MLabHubdjango/comment/admin.py
@@ -1,4 +1,3 @@
 from django.contrib import admin
-from lab.models import Lab
+
 # Register your models here.
-admin.site.register(Lab)
