+++ MLabHubdjango/lab/views.py
@@ -10,13 +10,35 @@ from rest_framework.response import Response
 from rest_framework import generics, permissions, status
 import pprint
 import json
-from api.models import Lab
+from lab.models import Lab
 from .serializers import LabSerializer
+from django.db.models import Q
+
 
 class GetLabInfo(APIView):
     permission_classes = (permissions.AllowAny, )
     def get(self, request):
-        labs = Lab.objects.all()
+        # only retrieve lab info that has been approved
+        labs = Lab.objects.filter(~Q(approved=False))
         labs = LabSerializer(labs, many = True)
         return Response(labs.data)
+    
+
+class GetDetailedLabInfo(generics.GenericAPIView):
+    permission_classes = (permissions.AllowAny, )
+    def get(self, request, id):
+        try:
+            lab = Lab.objects.get(pk=id)
+        except Lab.DoesNotExist:
+            return JsonResponse({"error": "Lab not found"}, status=404)
+        lab_data = {
+            "id": lab.id,
+            "name": lab.name,
+            "link": lab.link,
+            "intro": lab.intro,
+            "people": lab.people,
+        }
+        
+        return JsonResponse(lab_data, safe = False)
+
